+++ Package.swift
@@ -3,13 +3,8 @@ import PackageDescription
 
 let package = Package(
     name: "XCDYouTubeKit",
-    platforms: [
-        .iOS(.v8),
-        .tvOS(.v9),
-        .macOS(.v10_10)
-    ],
     products: [
-        .library(name: "XCDYouTubeKit" , targets: ["XCDYouTubeKit"])
+        .library(name: "XCDYouTubeKit", targets: ["XCDYouTubeKit"])
     ],
     targets: [
         .target(