+++ XCDYouTubeKit/XCDYouTubeVideoWebpage.m
@@ -6,11 +6,13 @@
 
 @interface XCDYouTubeVideoWebpage ()
 @property (nonatomic, readonly) NSString *html;
+@property (nonatomic, readonly) NSDictionary *playerContextConfiguration;
 @end
 
 @implementation XCDYouTubeVideoWebpage
 
 @synthesize playerConfiguration = _playerConfiguration;
+@synthesize playerContextConfiguration = _playerContextConfiguration;
 @synthesize videoInfo = _videoInfo;
 @synthesize sts = _sts;
 @synthesize javaScriptPlayerURL = _javaScriptPlayerURL;
@@ -31,29 +33,43 @@
 {
 	if (!_playerConfiguration)
 	{
-		__block NSDictionary *playerConfigurationDictionary;
-		NSRegularExpression *playerConfigRegularExpression = [NSRegularExpression regularExpressionWithPattern:@"ytplayer.config\\s*=\\s*(\\{.*?\\});|[\\({]\\s*'PLAYER_CONFIG'[,:]\\s*(\\{.*?\\})\\s*(?:,'|\\))" options:NSRegularExpressionCaseInsensitive error:NULL];
-		[playerConfigRegularExpression enumerateMatchesInString:self.html options:(NSMatchingOptions)0 range:NSMakeRange(0, self.html.length) usingBlock:^(NSTextCheckingResult *result, NSMatchingFlags flags, BOOL *stop)
-		{
-			for (NSUInteger i = 1; i < result.numberOfRanges; i++)
+		NSArray<NSString *>*patterns = @[@";ytplayer\\.config\\s*=\\s*(\\{.+?\\});ytplayer", @";ytplayer\\.config\\s*=\\s*(\\{.+?\\});"];
+		for (NSString *pattern in patterns) {
+			NSDictionary *configuration = XCDPlayerConfigurationWithString(self.html, pattern);
+			if (configuration != nil)
 			{
-				NSRange range = [result rangeAtIndex:i];
-				if (range.length == 0)
-					continue;
-				
-				NSString *configString = [self.html substringWithRange:range];
-				NSData *configData = [configString dataUsingEncoding:NSUTF8StringEncoding];
-				NSDictionary *playerConfiguration = [NSJSONSerialization JSONObjectWithData:configData ?: [NSData new] options:(NSJSONReadingOptions)0 error:NULL];
-				if ([playerConfiguration isKindOfClass:[NSDictionary class]])
-				{
-					playerConfigurationDictionary = playerConfiguration;
-					*stop = YES;
-				}
+				_playerConfiguration = configuration;
+				break;
 			}
-		}];
-		_playerConfiguration = playerConfigurationDictionary;
+		}
 	}
 	return _playerConfiguration;
+}
+
+static NSDictionary *XCDPlayerConfigurationWithString(NSString *html, NSString *regularExpressionPattern)
+{
+	__block NSDictionary *playerConfigurationDictionary;
+	NSRegularExpression *playerConfigRegularExpression = [NSRegularExpression regularExpressionWithPattern:regularExpressionPattern options:NSRegularExpressionCaseInsensitive error:NULL];
+	[playerConfigRegularExpression enumerateMatchesInString:html options:(NSMatchingOptions)0 range:NSMakeRange(0, html.length) usingBlock:^(NSTextCheckingResult *result, NSMatchingFlags flags, BOOL *stop)
+	{
+		for (NSUInteger i = 1; i < result.numberOfRanges; i++)
+		{
+			NSRange range = [result rangeAtIndex:i];
+			if (range.length == 0)
+				continue;
+			
+			NSString *configString = [html substringWithRange:range];
+			NSData *configData = [configString dataUsingEncoding:NSUTF8StringEncoding];
+			NSDictionary *playerConfiguration = [NSJSONSerialization JSONObjectWithData:configData ?: [NSData new] options:(NSJSONReadingOptions)0 error:NULL];
+			if ([playerConfiguration isKindOfClass:[NSDictionary class]])
+			{
+				playerConfigurationDictionary = playerConfiguration;
+				*stop = YES;
+			}
+		}
+	}];
+	
+	return  playerConfigurationDictionary;
 }
 
 - (NSDictionary *) videoInfo
@@ -61,6 +77,11 @@
 	if (!_videoInfo)
 	{
 		NSDictionary *args = self.playerConfiguration[@"args"];
+		if (args == nil)
+		{
+			_videoInfo = XCDPlayerConfigurationWithString(self.html, @"ytInitialPlayerResponse\\s*=\\s*(\\{.+?\\})\\s*;");
+			return _videoInfo;
+		}
 		if ([args isKindOfClass:[NSDictionary class]])
 		{
 			NSMutableDictionary *info = [NSMutableDictionary new];
@@ -99,6 +120,18 @@
 	
 	return _sts;
 }
+
+static NSURL *XCDJavaScriptPlayerURLFromString(NSString *javaScriptString)
+{
+	NSString *javaScriptPlayerURLString = javaScriptString;
+	if ([javaScriptString hasPrefix:@"//"])
+		javaScriptPlayerURLString = [@"https:" stringByAppendingString:javaScriptString];
+	else if ([javaScriptString hasPrefix:@"/"])
+		javaScriptPlayerURLString = [@"https://www.youtube.com" stringByAppendingString:javaScriptString];
+	
+	return [NSURL URLWithString:javaScriptPlayerURLString];
+}
+
 - (NSURL *) javaScriptPlayerURL
 {
 	if (!_javaScriptPlayerURL)
@@ -106,42 +139,35 @@
 		NSString *jsAssets = [self.playerConfiguration valueForKeyPath:@"assets.js"];
 		if ([jsAssets isKindOfClass:[NSString class]])
 		{
-			NSString *javaScriptPlayerURLString = jsAssets;
-			if ([jsAssets hasPrefix:@"//"])
-				javaScriptPlayerURLString = [@"https:" stringByAppendingString:jsAssets];
-			else if ([jsAssets hasPrefix:@"/"])
-				javaScriptPlayerURLString = [@"https://www.youtube.com" stringByAppendingString:jsAssets];
-			
-			_javaScriptPlayerURL = [NSURL URLWithString:javaScriptPlayerURLString];
+			_javaScriptPlayerURL = XCDJavaScriptPlayerURLFromString(jsAssets);
 		}
 		else
 		{
-			NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"\"assets\":.+?\"js\":\\s*(\"[^\"]+\")" options:0 error:nil];
-			NSTextCheckingResult *result = [regex firstMatchInString:self.html options:(NSMatchingOptions)0 range:NSMakeRange(0, self.html.length)];
-			if (result.numberOfRanges < 2)
-				return _javaScriptPlayerURL;
-			
-			NSRange range = [result rangeAtIndex:1];
-			if (range.length == 0)
-				return _javaScriptPlayerURL;
-			
-			
-			NSString *baseJSURLPathString = [[[self.html substringWithRange:range] stringByReplacingOccurrencesOfString:@"\\" withString:@""] stringByReplacingOccurrencesOfString:@"\"" withString:@""];
-			
-			NSURLComponents *components = [NSURLComponents componentsWithString:baseJSURLPathString];
-			if (components == nil)
-				return _javaScriptPlayerURL;
-			
-			if (components.scheme == nil || components.scheme.length == 0)
-				components.scheme = @"https";
-			
-			if (components.host == nil || components.host.length == 0)
-				components.host = @"www.youtube.com";
-			
-			if (components.URL == nil)
-				return _javaScriptPlayerURL;
-			
-			_javaScriptPlayerURL = components.URL;
+			NSArray<NSString *>*patterns = @[@"<script[^>]+\\bsrc=(\"[^\"]+\")[^>]+\\bname=[\"\\\']player_ias/base", @"jsUrl\"\\s*:\\s*(\"[^\"]+\")", @"\"assets\":.+?\"js\":\\s*(\"[^\"]+\")"];
+			for (NSString *pattern in patterns)
+			{
+				NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:pattern options:0 error:nil];
+				NSTextCheckingResult *result = [regex firstMatchInString:self.html options:(NSMatchingOptions)0 range:NSMakeRange(0, self.html.length)];
+				if (result.numberOfRanges < 2)
+				{
+					continue;
+				}
+				
+				NSRange range = [result rangeAtIndex:1];
+				if (range.length == 0)
+				{
+					continue;
+				}
+				
+				NSString *baseURLString = [[[self.html substringWithRange:range] stringByReplacingOccurrencesOfString:@"\"" withString:@""] stringByReplacingOccurrencesOfString:@"\\" withString:@""];
+				NSURL *javaScriptPlayerURL = XCDJavaScriptPlayerURLFromString(baseURLString);
+				if (javaScriptPlayerURL != nil)
+				{
+					_javaScriptPlayerURL = javaScriptPlayerURL;
+					break;
+				}
+				
+			}
 		}
 	}
 	return _javaScriptPlayerURL;